refactor: extract component SearchLinkSection
